SPOI-5389 - Support for RM delegation token renewal in secure HA environments
