+++ engine/src/main/java/com/datatorrent/stram/StramClient.java
@@ -18,13 +18,11 @@ package com.datatorrent.stram;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.net.InetSocketAddress;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.nio.ByteBuffer;
 import java.util.*;
 
-import com.google.common.base.Joiner;
 import com.google.common.base.Objects;
 import com.google.common.collect.Lists;
 
@@ -39,10 +37,7 @@ import org.apache.hadoop.classification.InterfaceStability;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.*;
 import org.apache.hadoop.io.DataOutputBuffer;
-import org.apache.hadoop.io.Text;
-import org.apache.hadoop.net.NetUtils;
 import org.apache.hadoop.security.Credentials;
-import org.apache.hadoop.security.SecurityUtil;
 import org.apache.hadoop.security.UserGroupInformation;
 import org.apache.hadoop.security.token.Token;
 import org.apache.hadoop.util.JarFinder;
@@ -52,7 +47,6 @@ import org.apache.hadoop.yarn.client.api.YarnClient;
 import org.apache.hadoop.yarn.client.api.YarnClientApplication;
 import org.apache.hadoop.yarn.conf.YarnConfiguration;
 import org.apache.hadoop.yarn.exceptions.YarnException;
-import org.apache.hadoop.yarn.security.client.RMDelegationTokenIdentifier;
 import org.apache.hadoop.yarn.util.ConverterUtils;
 import org.apache.hadoop.yarn.util.Records;
 import org.apache.log4j.DTLoggerFactory;
@@ -79,13 +73,6 @@ public class StramClient
   private static final Logger LOG = LoggerFactory.getLogger(StramClient.class);
   public static final String YARN_APPLICATION_TYPE = "DataTorrent";
   public static final String LIB_JARS_SEP = ",";
-
-  // TODO: HADOOP UPGRADE - replace with YarnConfiguration constants
-  private static final String RM_HA_PREFIX = YarnConfiguration.RM_PREFIX + "ha.";
-  private static final String RM_HA_IDS = RM_HA_PREFIX + "rm-ids";
-  private static final String RM_HA_ENABLED = RM_HA_PREFIX + "enabled";
-  private static final boolean DEFAULT_RM_HA_ENABLED = false;
-  private static final String RM_HOSTNAME_PREFIX = YarnConfiguration.RM_PREFIX + "hostname.";
 
   // Configuration
   private final Configuration conf;
@@ -300,51 +287,6 @@ public class StramClient
 
   }
 
-  // TODO: HADOOP UPGRADE - replace with YarnConfiguration constants
-  private Token<RMDelegationTokenIdentifier> getRMHAToken(org.apache.hadoop.yarn.api.records.Token rmDelegationToken) {
-    // Build a list of service addresses to form the service name
-    ArrayList<String> services = new ArrayList<String>();
-    for (String rmId : conf.getStringCollection(RM_HA_IDS)) {
-      LOG.info("Yarn Resource Manager id: {}", rmId);
-      // Set RM_ID to get the corresponding RM_ADDRESS
-      services.add(SecurityUtil.buildTokenService(NetUtils.createSocketAddr(
-              conf.get(RM_HOSTNAME_PREFIX + rmId),
-              YarnConfiguration.DEFAULT_RM_PORT,
-              RM_HOSTNAME_PREFIX + rmId)).toString());
-    }
-    Text rmTokenService = new Text(Joiner.on(',').join(services));
-
-    return new Token<RMDelegationTokenIdentifier>(
-            rmDelegationToken.getIdentifier().array(),
-            rmDelegationToken.getPassword().array(),
-            new Text(rmDelegationToken.getKind()),
-            rmTokenService);
-  }
-
-  private void addRMDelegationToken(final String renewer, final Credentials credentials) throws IOException, YarnException {
-    // Get the ResourceManager delegation rmToken
-    final org.apache.hadoop.yarn.api.records.Token rmDelegationToken = yarnClient.getRMDelegationToken(new Text(renewer));
-
-    Token<RMDelegationTokenIdentifier> token;
-    // TODO: Use the utility method getRMDelegationTokenService in ClientRMProxy to remove the separate handling of
-    // TODO: HA and non-HA cases when hadoop dependency is changed to hadoop 2.4 or above
-    if (conf.getBoolean(RM_HA_ENABLED, DEFAULT_RM_HA_ENABLED)) {
-      LOG.info("Yarn Resource Manager HA is enabled");
-      token = getRMHAToken(rmDelegationToken);
-    } else {
-      LOG.info("Yarn Resource Manager HA is not enabled");
-      InetSocketAddress rmAddress = conf.getSocketAddr(YarnConfiguration.RM_ADDRESS,
-              YarnConfiguration.DEFAULT_RM_ADDRESS,
-              YarnConfiguration.DEFAULT_RM_PORT);
-
-      token = ConverterUtils.convertFromYarn(rmDelegationToken, rmAddress);
-    }
-
-    LOG.info("RM dt {}", token);
-
-    credentials.addToken(token.getService(), token);
-  }
-
   /**
    * Launch application for the dag represented by this client.
    *
@@ -461,7 +403,7 @@ public class StramClient
         fs.close();
       }
 
-      addRMDelegationToken(tokenRenewer, credentials);
+      new ClientRMHelper(yarnClient, conf).addRMDelegationToken(tokenRenewer, credentials);
 
       DataOutputBuffer dob = new DataOutputBuffer();
       credentials.writeTokenStorageToStream(dob);
@@ -718,7 +660,7 @@ public class StramClient
       }
 
     };
-    ClientRMHelper rmClient = new ClientRMHelper(yarnClient);
+    ClientRMHelper rmClient = new ClientRMHelper(yarnClient, conf);
     return rmClient.waitForCompletion(appId, callback, clientTimeout);
   }
